clean up `test_no_imports` tests
